+++ Csharp/Anagrams.cs
