+++ Csharp/PythagoreanTriplet.cs
